# docs/swagger.yaml
basePath: /api
...
host: localhost:8080
info:
  contact: {}
  description: API для расчета масс звезд в заявках
  title: Stellar Mass Calculation API
  version: "1.0"
paths:
...
  /mass-requests/{id}/classes/{class_id}:
    put:
      consumes:
      - application/json
      description: Update class parameters in mass request
      parameters:
      - description: Request ID
        in: path
        name: request_id
        required: true
        type: integer
      - description: Class ID
        in: path
        name: class_id
        required: true
        type: integer
      - description: Update data
        in: body
        name: updates
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update request class item
      tags:
      - request-classes
...
  /mass-requests/{request_id}/classes/{class_id}:
    delete:
      consumes:
      - application/json
      description: Remove spectral class from mass request
      parameters:
      - description: Request ID
        in: path
        name: request_id
        required: true
        type: integer
      - description: Class ID
        in: path
        name: class_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Remove class from request
      tags:
      - request-classes
  /mass-requests/classes:
    post:
      consumes:
      - application/json
      description: Add spectral class to draft mass request
      parameters:
      - description: Class data
        in: body
        name: request
        required: true
        schema:
          additionalProperties:
            format: int64
            type: integer
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Add class to request
      tags:
      - request-classes
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"